Simple registrar hardhat tests
This feature is only for reference. I do not see much added value of merging this
as these tests are already covered in cljs files.

However, the JS hardhat tests are easier to debug, because
* You can easily run just a single test (e.g. `it.only`)
* You can use hardhat debugging tools (console.log and also built in **stack traces**)
* The tests are faster

It might be useful in the future, if we have a hard-to-debug bug in contracts.

You can run the tests by: `npx hardhat test`. It will compile the contracts and spawn hardhat node (alternative to ganache).

P.S. I also tried replacing our `compile-solidity.sh` with hardhat compile step `npx hardhat compile`, but there were some issues:
* https://github.com/nomiclabs/hardhat/issues/959
* In dev mode, we postprocess the compiled contracts and replace the generated `bin` files. This is not really compatible with the way how hardhat compile works.